Manual fixing of duplicates
The duplicates were caused (one could argue fixed) because oragono allows 25-character usernames instead of 9-character ones.

Of course, since the actual tildescores are different right now, this file is not up to date, so all this effort was for nothing.

A real solution would be to edit tildebot.py so that it gracefully updated old 9-character matches properly.

Another solution is to make tildebot quit, make the required changes to tildescores.txt, then make tildebot join again.